conv: Use gate kernel object + improve visualization of excluded threads
If thread A switches to a tracing excluded thread X which switches to
excluded Y and back to A this now shows A->X X->A instead of A->X Y->A.
Before, X would appear to be running indefinitely from this point
onward. Now Ys runtime goes missing, but you can't rely on the runtime
of the excluded threads anyway, so this is a clearer solution.